Created RedDeerException. Fixes #574
