+++ plugins/org.jboss.reddeer.jface/src/org/jboss/reddeer/eclipse/jface/exception/JFaceLayerException.java
@@ -1,12 +1,14 @@
 package org.jboss.reddeer.eclipse.jface.exception;
 
+import org.jboss.reddeer.swt.exception.RedDeerException;
+
 /**
  * Thrown when an error can be identified on the JFace layer
  * 
  * @author Lucia Jelinkova
  *
  */
-public class JFaceLayerException extends RuntimeException {
+public class JFaceLayerException extends RedDeerException {
 
 	private static final long serialVersionUID = -3641202955039921211L;
