+++ plugins/org.jboss.reddeer.requirements/META-INF/MANIFEST.MF
@@ -6,9 +6,10 @@ Bundle-Version: 0.6.0.qualifier
 Bundle-Activator: org.jboss.reddeer.requirements.Activator
 Require-Bundle: org.eclipse.ui,
  org.eclipse.core.runtime,
- org.jboss.reddeer.junit;bundle-version="0.4.0",
- org.jboss.reddeer.eclipse;bundle-version="0.4.0",
- org.jboss.reddeer.workbench;bundle-version="0.4.0"
+ org.jboss.reddeer.junit,
+ org.jboss.reddeer.eclipse,
+ org.jboss.reddeer.workbench,
+ org.jboss.reddeer.swt
 Bundle-RequiredExecutionEnvironment: JavaSE-1.6
 Bundle-ActivationPolicy: lazy
 Bundle-Vendor: JBoss by Red Hat
